+++ sqlparser/test/parser/moor_file_test.dart
@@ -15,7 +15,7 @@ CREATE TABLE tbl (
 
 all: SELECT /* COUNT(*), */ * FROM tbl WHERE $predicate;
 @special: SELECT * FROM tbl;
-typeHints(:foo AS TEXT): SELECT :foo;
+typeHints(:foo AS TEXT, $predicate = TRUE): SELECT :foo WHERE $predicate;
 nested AS MyResultSet: SELECT foo.** FROM tbl foo;
 ''';
 
@@ -85,7 +85,11 @@ void main() {
                 ColonVariableToken(fakeSpan(':foo'), ':foo'),
               ),
               'TEXT',
-            )
+            ),
+            DartPlaceholderDefaultValue(
+              'predicate',
+              BooleanLiteral.withTrue(token(TokenType.$true)),
+            ),
           ],
         ),
         DeclaredStatement(